Finalized

+++ Gestion/Server/Services/GoalService.cs
@@ -19,10 +19,11 @@ public class GoalService
         return goals.Select(goal => new GoalDTO
         {
             GoalId = goal.GoalId,
-            GoalName = goal.GoalName,
+            Name = goal.GoalName,
             CreatedDate = goal.CreatedDate,
             CompletedTasks = goal.CompletedTasks,
-            TotalTasks = goal.TotalTasks
+            TotalTasks = goal.TotalTasks,
+            Percentage= goal.Percentage,
         })
         .ToList(); 
     }
@@ -34,67 +35,64 @@ public class GoalService
         return new GoalDTO
         {
             GoalId = goal.GoalId,
-            GoalName = goal.GoalName,
+            Name = goal.GoalName,
             CreatedDate = goal.CreatedDate,
             CompletedTasks = goal.CompletedTasks,
-            TotalTasks = goal.TotalTasks
+            TotalTasks = goal.TotalTasks,
+            Percentage= goal.Percentage,
         };
     }
     public async Task Save(GoalUpdateDTO goal)
     {
         var search = await _context.Goals
-                                 .FirstOrDefaultAsync(g => g.GoalName == goal.GoalName.Trim());
+                                 .FirstOrDefaultAsync(g => g.GoalName == goal.Name.Trim());
         if(search != null)
         {
             throw new Exception("El nombre no esta disponible.");
         }
         Goal _goal = new()
         {
-            
-            GoalName = goal.GoalName,
+
+            GoalName = goal.Name,
             CreatedDate = DateTime.Now,
             CompletedTasks = 0,
-            TotalTasks = 0
+            TotalTasks = 0,
+            Percentage = 0
         };
         _context.Goals.Add(_goal );
         await _context.SaveChangesAsync();
     }
     public async Task Update(int goalId, GoalUpdateDTO updatedGoal)
-    {
-        // Busca el objetivo por su ID
+    { 
         var goal = await _context.Goals.FirstOrDefaultAsync(g => g.GoalId == goalId);
 
         if (goal == null)
         {
-            throw new Exception("El objetivo no existe.");
-        }
-
-        // Verifica si el nombre del objetivo ya existe (excepto para el objetivo actual)
+            throw new Exception("La meta no existe.");
+        } 
         var search = await _context.Goals
-                                   .FirstOrDefaultAsync(g => g.GoalName == updatedGoal.GoalName.Trim() && g.GoalId != goalId);
+                                   .FirstOrDefaultAsync(g => g.GoalName == updatedGoal.Name.Trim() && g.GoalId != goalId);
 
         if (search != null)
         {
             throw new Exception("El nombre no está disponible.");
-        } 
-        // Actualiza los campos necesarios
-        goal.GoalName = updatedGoal.GoalName; 
-
-        // Guarda los cambios en la base de datos
+        }  
+        goal.GoalName = updatedGoal.Name;  
         _context.Goals.Update(goal);
         await _context.SaveChangesAsync();
     }
     public async Task Delete(int goalId)
-    {
-        // Busca el objetivo por su ID
+    { 
         var goal = await _context.Goals.FirstOrDefaultAsync(g => g.GoalId == goalId);
-
         if (goal == null)
         {
             throw new Exception("El objetivo no existe.");
         }
-
-        // Elimina el objetivo
+        var hasAssignments = await _context.Assignments.AnyAsync(a => a.GoalId == goalId);
+        if (hasAssignments)
+        {
+            throw new Exception("No se puede eliminar la meta porque tiene tareas asignadas.");
+        } 
         _context.Goals.Remove(goal);
         await _context.SaveChangesAsync();
     }
